refactor code

+++ backend/src/lambda/http/deleteTodo.ts
@@ -3,8 +3,8 @@ import 'source-map-support/register'
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import * as middy from 'middy'
 import { cors, httpErrorHandler } from 'middy/middlewares'
-
-import { deleteTodo } from '../../businessLogic/todos'
+import { http_response } from '../../common/http-response'
+import { deleteTodo } from '../../helpers/todos'
 import { getUserId } from '../utils'
 import { createLogger } from '../../utils/logger'
 
@@ -15,13 +15,16 @@ export const handler = middy(
     const todoId = event.pathParameters.todoId
     // TODO: Remove a TODO item by id
     const userId = getUserId(event)
-    logger.info('Deleting todo item', { todoId, userId })
+    if (!userId) return http_response._404({ error: 'Not found' })
+    logger.info('To delete todo', { todoId, userId })
     const res = await deleteTodo(todoId, userId)
-    logger.info('Deleted todo item', { todoId, userId, res })
-    return {
-      statusCode: 200,
-      body: JSON.stringify({})
+    logger.info('Deleted todo', { todoId, userId, res })
+    if (!res) {
+      return http_response._400({ error: 'Unable to delete the user' })
     }
+    return http_response._200({
+      message: `Todo with ID; ${todoId} is deleted!`
+    })
   }
 )
